Fixing dark mode text color
