+++ src/app/layout.tsx
@@ -1,13 +1,9 @@
-import type { Metadata } from "next";
-import type { ReactNode } from "react";
+import { ColorSchemeScript, mantineHtmlProps } from "@mantine/core";
 import "@mantine/core/styles.css";
-import {
-  createTheme,
-  ColorSchemeScript,
-  MantineProvider,
-  mantineHtmlProps,
-} from "@mantine/core";
+import type { Metadata } from "next";
 import { Roboto } from "next/font/google";
+import type { ReactNode } from "react";
+import ThemeProvider from "./components/theme";
 
 const roboto = Roboto({
   subsets: ["latin"],
@@ -17,13 +13,6 @@ export const metadata: Metadata = {
   description: "Generated by create next app",
   title: "Bryan Laporte | Profile",
 };
-
-const theme = createTheme({
-  fontFamily: "Roboto, Arial, Verdana, sans-serif",
-  headings: {
-    fontWeight: "500",
-  },
-});
 
 const Layout = ({
   children,
@@ -39,9 +28,7 @@ const Layout = ({
         />
       </head>
       <body>
-        <MantineProvider defaultColorScheme="auto" theme={theme}>
-          {children}
-        </MantineProvider>
+        <ThemeProvider>{children}</ThemeProvider>
       </body>
     </html>
   );
